+++ vcontrold/unit.h
@@ -1,6 +1,6 @@
 /* unit.h */
 /* $Id$ */
 
-int procGetUnit(unitPtr uPtr, char *recvBuf, int len, char *result, char bitpos, char *pRecvPtr);
-int procSetUnit(unitPtr uPtr, char *sendBuf, short *sendLen, char bitpos, char *pRecvPtr);
+int procGetUnit(unitPtr uPtr, char *recvBuf, size_t len, char *result, char bitpos, char *pRecvPtr);
+int procSetUnit(unitPtr uPtr, char *sendBuf, size_t *sendLen, char bitpos, char *pRecvPtr);
